Task # 51, 52, 53, 54, 60, 66
Task # 51 |  Time taken : 1 ½  hours
Made code changes under all pages to reflect the new copyright info. Results on the simulator are in line with the customer requirements.

Task # 52, 53, 54  |  Time taken : 1 hour
Modifications to the code under "CreditDetails" page to reflect the new content. Results on the simulator are in line with the customer requirements.

Task # 66  |  Time taken : 1 hour
Downloaded and updated the latest official logos from the corresponding networking and media content apps as per the norms declared in their corresponding websites, across Connect and Gallery pages.

Task # 60  |  Time taken : 2 hours
Changes had to be made to the source code of the “storyboard” that are used to build the interfaces in the iOS app.