Task # 68, 69  |  Time taken : 3 ½ hours
Task # 68, 69  |  Time taken : 3 ½ hours
The “News + Events” page was modified to include the bar titled "In the Media" just below "News". This was successfully completed and tested. Clicking on this “In the Media” bar now takes the user to “https://iho.asu.edu/media”. The News link in the "News + Events" page has been modified as well. This involved modifying the main storyboard of the iOS app. The flow in the storyboard has been altered such that clicking on the News now takes the user to IHO News page "https://iho.asu.edu/news". Now, in this iOS app, the News bar does not direct user to the News page of the app, but instead it goes to the URL mentioned above.