Task # 75  |  Time taken : 4 ½  hours
The “Travel + Learn” page was modified to have a picture that will look like a banner. A lot of efforts had to be pitched to understand and modify the picture to the specifications that the customer needs. The UX of the iOS app had to be tweaked to get the intended results where the image displays fine in screen sizes of iphone 8 and X. Other relevant code related modifications required for the “Travel + Learn” page was also carried out.