Task # 79, 82, 83  |  Time taken : 4 ½ hours
In Android app, the “News + Events” page was modified to include the bar titled "In the Media" just below "News". This was successfully completed and tested. Clicking on this “In the Media” bar now takes the user to “https://iho.asu.edu/media”. Next, the “Travel + Learn” page was modified to have a picture that will look like a banner. A lot of efforts had to be pitched to understand and modify the picture to the specifications that the customer needs. The UX design of the android app had to be tweaked to get the intended results. Other relevant code related modifications required for the “Travel + Learn” page was also carried out.