Task # 109  |  Time taken : 1 hour
Work in Progress. Unlike before for similar functionality (Task # 68, 69), for some reason, the changes made to the code and Main storyboard are not being reflected when I run the simulation of the app. I have followed the same steps that I had for similar deployments in Sprint 2, but the changes are not being reflected when i try today. Debugging the app to see what is going wrong.