Task # 109, 140, 114, 138, 139, 136, 137
Task # 109, 140  |  Time taken : 2 hours
For task 109, I debugged the code in Main storyboard of the iOS app to fix the broken connections between the UI and code components. Post implementing this fix, the bars work correctly, as intended. Similar approach was followed task 140.

Task # 114, 138, 139, 136, 137  |  Time taken : 2 hours
In Android “Roboto” font which is based on “San-Serif”, is the default font used. However, in iOS, this font is not available even for selection. Hence, I have imported the Roboto font into the “resources” of folder of iOS project, and then made configurational changes to the iOS app to include the newly imported font. Next, I am going page by page to change font style to Roboto and increasing the font size to a bigger value (19 for headings and 14 for normal text) so that the text is easily readable unlike the status quo where the text is a little too small.