Task # 114, 120, 134, 135, 136, 137, 138, 152, 155, 156, 157, 159, 158, 160  |  Time taken : 11 hours 25 minutes
Task # 134, 135  |  Time taken : 30 minutes
In iOS app, the "Events" bar in “News + Events” page was modified to take the user to the IHO Events page via Safari browser. This has been successfully completed. This involved modifying the main storyboard of the iOS app. The flow in the storyboard has been altered such that clicking on the Events now takes the user to IHO News page "https://iho.asu.edu/upcoming-events". Now, in this iOS app, the Events bar does not direct user to the Events page of the app, but instead it goes to the URL mentioned above. Similarly, the "Sign up for E-News" bar in “Connect” page was modified to take the user to the IHO Subscribe page via the mobile Safari browser.

Task # 120  |  Time taken : 4 ½ hours
Work in Progress. Tried a lot of fixes to resolve this page scrolling issue in iOS app. But, none of them works. It appears, a new page has to be created and designed, to replace the current home page. So, will put a pause on this task for now and continue with other tasks since this task has the lowest priority as per customer's revised requirements. Regarding the priority of this task, this info is available at the beginning of the document titled “IHO App review REVISED 10-26-2018.docx”, under the heading "Priorities for November to finish the app". This document has uploaded to Taiga under this task, as an attachment.

Task # 152  |  Time taken : 20 minutes
In iOS app, In “Becoming Human” page, the buttons were converted into bars. They work as intended across varied screen sizes.

Task # 136, 137, 159, 160  |  Time taken : 45 minutes
In iOS app, For Homepage, there are 6 buttons. For “News + Events” page, there are just 5 bars. For “Connect” page, there are just 4 bars. For “Field Notes” page there are 4 bars, with the sub-text in the fourth bar named “Blog-Notes from the Field” to have a font size of 15 only. The Font style for these bars and buttons in the respective pages were changed to Roboto and Font size was increased to 19, and 15 for the sub-text mentioned above.

Task # 114, 155, 158  |  Time taken : 2 ½ hours
In iOS app, For pages “Featured News”, “Travel + Learn” and “Becoming Human”, they have a heading, followed by an image, then a short passage / article, and a button which takes the user to the IHO webpage via Safari browser. Change Font style of all text in "Featured News" page to Roboto. Change Font size of the Heading to size 19. Change Font size of the article to size 16. Change Font size of button to size 19. Here, changing the font sizes threw all the components in these pages completely out of alignment due to prior poor design implementation. The page designs were redone using appropriate constraints so that the design remains intact across multiple screen sizes. Hence, this took a lot of time & effort coupled with repeated testings for every change made !

Task # 138  |  Time taken : 1 hour
In iOS app, For the landing page of “Gallery”, the font style was changed to Roboto and size to 19 for the button present in this page. In this page, the text present above "Youtube" and "Vimeo" buttons was changed to "Follow us on the following channels" with its font style as roboto and rise at 17.
Again here, changing the font size for the text threw all the components in this page completely out of alignment due to prior poor design implementation. The page designs were redone using appropriate constraints so that the design remains intact across multiple screen sizes. Hence, this too took a lot of time & effort coupled with repeated testings for every change made !!

Task # 156 |  Time taken : 1 ½ hours
In iOS app, For “Ask an Anthropologist” page, this pages has an image, then a short passage / article, and a button which takes the user to the webpage "https://askananthropologist.asu.edu" via Safari browser. Font Style of button was changed to Roboto, and font size was changed to 19. Spent a lot of time trying to change the font style and font size of the article in this page. The text of the article in this page is extracted from a webpage using UIWebView components. Currently, Swift does NOT support font attributes manipulations for the text extracted using UIWebView class, unlike UITextView. However, UITextView has complications when it comes to data extraction from a webpage. Hence, the text in the article of this page will continue to remain as-is due to Swift limitations.

Task # 157 |  Time taken : 20 minutes
In iOS app, IHO logo constraints have been updated in the homepage so that the complete logo is displayed irrespective of the screen sizes, without any edges being chipped off.