Task # 120, 151  |  Time taken : 6 ½  hours
In iOS app, the pages such as Homepage, "Ask an Anthropologist", "Becoming Human", "Featured News", and others are not scrollable because they were built using plain UIView modules.  For the pages to be scrollable, multi-screen sizes compatible, and Portrait & Landscape compatible, these pages need to be re-built from the scratch, using a combination of the UIViewController, UIScrollView and UIContentView modules along with appropriate auto-layout constraints. There is a lot of work involved here and it would involve a lot of time to carry these out, especially since they need to be tested well.
The resolution has been implemented for the homepage of the app, and it works fine as it should have ideally been. Considering the amount of time spent and the effort pitched in, I am quite satisfied with the outcome and looking forward to implementing the same resolution for other pages.
I also re-organized the Main storyboard of the iOS app, after removing components that are no longer needed in the app.