Task # 165, 166  |  Time taken : 2 ½  hours
In iOS app, to solve the scrolling issue and design discrepancy when screen orientation changes from Portrait to Landscape, the pages "Ask an Anthropologist" and "Becoming Human" were re-built using UIViewController, UIScrollView, UIContentView and appropriate auto-layout constraints so that the pages are scrollable and resizable across multiple screen sizes and screen orientations. Irrespective of the screen size or orientation, the design and layout of the pages now remain consistent throughout.