Task # 161, 167, 168, 180, 181  |  Time taken : 4 hours 30 mins
Task # 161, 180  |  Time taken : 30 minutes
In iOS app, the font style of all the text in pages “Who is Lucy” and “About”, were updated to roboto. The headings in these pages were increased to size 20, and general text size were increased to 17.

Task # 167, 168  |  Time taken : 3 hours
In iOS app, the Gallery landing page and the About page were completely rebuilt to fix the scrolling and alignment issues. I tried to reuse the components already built, in the new pages that I had created, but that only ended up in added debugging efforts ! .. As a result, I got rid of those old components and built new ones. Hence the reason for additional time taken compared to tasks 165 and 166.

Task # 181  |  Time taken : 1 hour
In “Who is Lucy” page, the pinch-to-zoom function did not work here. Even though this functionality is not officially requested in the requirements given by the customer, during one of the meetings the customer did casually mention that having this feature would be quite useful. Moreover, as a user, I found that not having this functionality was quite inconvenient. So, I tried to fix the zoom issue by altering the settings of the UIWebView component in this page. My efforts paid off and I was able to resolve this issue. Since there is no user story for this, I have created this task under “Unassigned Tasks” in Taiga board.