Task # 162, 169, 177  |  Time taken : 2 ½  hours
In iOS app, the Font Style of the text in the “Donate” page has been changed to Roboto. The font size of general text in this page has been changed to size 16 and that of the headings to 19. The “invest in ASU” logo currently in use is very low in resolution. Hence, there is a limit to its size increase. As a result, I have converted the logo button to a bar since the client prefers bars over buttons in every page. The colors used in the bar are as per ASU color specifications. This “Donate” page has been re-built using UIViewController, UIScrollView, UIContentView and appropriate auto-layout constraints to solve the scrolling issue and design discrepancy when screen orientation changes from Portrait to Landscape. The page is now scrollable and resizable across multiple screen sizes and screen orientations. Irrespective of the screen size or orientation, the design and layout of the page now remain consistent throughout.