Task # 182  |  Time taken : 4 ½ hours
For external URLs in the mobile apps, Client had given us the liberty to redirect users to either to a seperate browser or have it as in-app. Since the web-pages maintained by the client are in drupal servers, it was decided to use external browsers for rendering the exact content of the external URLs. However, this lead to too many “jumps” between IHO app and the web-browser. As a user, I was not happy with this experience. Hence, I took efforts to enable External URLs in the iOS app to be opened as in-app pages instead of redirecting the user to a safari browser. This involved creating 9 new pages, i.e. View Controllers, from the ground up, so that the web content is displayed appropriately as in-app content.
Following are the new pages created:
1. News
2. In the Media
3. Events
4. More on Travel
5. Blog Notes
6. Becoming Human
7. Ask an Anthropologist
8. Official IHO website from Connect page
9. Contact us from Connect page

Since there is no User story created for this task, I have created this task under Unassigned Tasks.
Please note that I do not want force this option on the client. Rather, I want to offer it as an option. So, I have forked a new branch named “VimalMenonSprint3-inAppWeb” from the VimalMenonSprint3 branch. Except for this task 182, the above mentioned two branches are the same.